refactor(HMI para test de CV)
Refactorizando la interfaz para la configuración y calibración de la webcam.
Renombre el script y añadí una clase de "WebCamParameters" que servirá como contenedor (como un puntero a un struct en c) guardando la información para acceder al la webcam mediante openCV, esto con el fin de poder hacer varias clases de objetos diferentes que sirvan como bloques para construir la interfaz completa y poder decidir que bloques de la interfaz necesita acceso a la webcam para realizar sus funciones o modificar parámetros.
